[filter/purge.c] get rid of compiler warning on 32-bit machine (hopefully)
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado-nfs/trunk@915 3eaf19af-ecc0-40f6-b43f-672aa0c71c71

+++ filter/purge.c
@@ -744,7 +744,7 @@ usage (void)
 static int
 approx_phi (long B)
 {
-  ASSERT_ALWAYS(B <= 53030236260); /* otherwise B/log(B) > 2^31 */
+  ASSERT_ALWAYS((double) B <= 53030236260.0); /* otherwise B/log(B) > 2^31 */
   return (B <= 1) ? 0 : (int) ((double) B / log ((double) B));
 }
